cleanups to allow compilation on BSD system
git-svn-id: svn+ssh://svn.it.su.se/svn/spocp/trunk@397 64b275d1-5406-0410-b6a0-e042953264b1

+++ lib/db0.c
@@ -1000,13 +1000,13 @@ ruleinst_print(ruleinst_t * r, char *rs)
 
 	octcat(oct, r->uid, 40);
 
-	snprintf(flen, 1024, "%d:", r->rule->len);
+	snprintf(flen, 1024, "%d:", (int) r->rule->len);
 	octcat(oct, flen, strlen(flen));
 
 	octcat(oct, r->rule->val, r->rule->len);
 
 	if (r->blob && r->blob->len) {
-		snprintf(flen, 1024, "%d:", r->blob->len);
+		snprintf(flen, 1024, "%d:", (int) r->blob->len);
 		octcat(oct, flen, strlen(flen));
 		octcat(oct, r->blob->val, r->blob->len);
 	}
